requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coverage==5.1
fastapi==0.55.1
firebase-admin==4.3.0
google-api-core==1.17.0
google-api-python-client==1.8.4
google-auth==1.16.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
h11==0.9.0
httplib2==0.18.1
idna==2.9
more-itertools==8.3.0
msgpack==1.0.0
packaging==20.4
pip-autoremove==0.9.1
pluggy==0.13.1
protobuf==3.12.2
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.5.1
pyparsing==2.4.7
pytest==5.4.2
pytest-cov==2.9.0
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
starlette==0.13.2
uritemplate==3.0.1
urllib3==1.25.9
uvicorn==0.11.5
wcwidth==0.1.9
websockets==8.1
